requirements.txt
fastapi==0.104.1
pyjwt==2.8.0
argon2-cffi==23.1.0
pymongo==4.9.0
python-dotenv==1.0.0
uvicorn==0.24.0
motor==3.6.0
email-validator==2.2.0
razorpay==1.4.1
replicate==0.22.0
python-jose==3.3.0
cryptography==41.0.5
pydantic_settings==2.7.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.7.0
typing_extensions==4.8.0
starlette==0.27.0
bcrypt==4.0.1
dnspython==2.4.2
requests==2.31.0
setuptools>=65.5.1
